adding verification for registering user by email and adding status for the profile and other changes

+++ src/main/java/com/menainnovations/weshare/services/PhotoServiceImpl.java
@@ -20,6 +20,7 @@ import java.util.List;
 @Service
 public class PhotoServiceImpl implements PhotoService {
     public static String ROOT = "uploads/images/post";
+    public static String ROOT2 = "uploads/images/profile";
     long count = 0;
     private final ResourceLoader resourceLoader;
 
@@ -32,6 +33,8 @@ public class PhotoServiceImpl implements PhotoService {
     PhotoRepository photoRepository;
     @Autowired
     PostRepository postRepository;
+    @Autowired
+    UserService userService;
 
 
     @Override
@@ -50,7 +53,7 @@ public class PhotoServiceImpl implements PhotoService {
                     if (!destination.exists()){
                         destination.mkdir();
                     }
-                    ImageIO.write(src, "png", destination);
+                    ImageIO.write(src, "jpg", destination);
                     count++;
                     /*
                         create a new Photo Obj to store info in DB photo table
@@ -68,12 +71,45 @@ public class PhotoServiceImpl implements PhotoService {
         }
         return "success";
     }
+    public String addProfilePhoto(long userId ,MultipartFile photo) {
+        File directory = new File(ROOT2 + "/" + userId);
+        if (!directory.exists()) {
+            directory.mkdir();
+        }
+        String photoName = new Date().toString().replace(" ", "").replace(":", "_");
+        if (!photo.isEmpty()) {
+            try {
+                BufferedImage src = ImageIO.read(new ByteArrayInputStream(photo.getBytes()));
+                File destination = new File(ROOT2 + "/" + userId + "/" + photoName + "22.jpg");
+                if (!destination.exists()) {
+                    destination.mkdir();
+                }
+                ImageIO.write(src, "jpg", destination);
+                    /*
+                        update user object to set profile pic
+                     */
+                    userService.updateUserProfilePhoto(userId,destination.toString().replace("\\","/"));
+
+            } catch (IOException | RuntimeException e) {
+                return "fail";
+            }
+        }
+        return "success";
+    }
+
 
     @Override
-    public void deletePhotoById(long id) {
-        photoRepository.delete(id);
-
-
+    public String deletePhotoById(long id) {
+        try {
+            Photo photo =photoRepository.findOne(id);
+            File file=new File(photo.getUrl());
+            if(file.delete()) {
+                photoRepository.delete(id);
+            }
+        }catch (Exception e){
+            return "fail";
+        }
+        return "success";
     }
 
     @Override
